Updated Command usage If you now want to use the Itemlocator via GUI or via command, you need to use the new Commands: `/itemlocator gui` - Open the GUI for tracking items `/itemlocator locate <material>` - Will execute the tracking of the Item
