+++ src/main/resources/lang/en.yml
@@ -7,3 +7,5 @@ no_update_available: "No updates available."
 player_only_command: "This command can only be executed by a player."
 found_item: "Found ${total} items of type ${item} at:"
 no_item_found: "No items of type ${item} found."
+usage_locate_command: "Wrong usage of command. Please use /itemlocator locate <item_name>"
+material_not_found: "The material '${material}' is not a valid Minecraft material."
